requirements.txt
django==2.0.4
django-crispy-forms==1.7.2
django-allauth==0.35.0
requests==2.18.4
talisker==0.9.6
psycopg2-binary==2.7.4
Pillow==5.0.0
